Rename LICENSE to @SKEEZY
"I'm a curious explorer with a passion for learning new things. I believe that life is an adventure, and I'm always looking for new challenges to tackle. My interests range from technology and programming to any platform. I'm always open to new ideas, and I love to engage in thoughtful conversations. If you're looking for a friendly, curious, and open-minded person to chat with, then I'm your guy!"

💡💡📝📍📍